+++ app/Room/models.py
@@ -1,7 +1,6 @@
-from decimal import Decimal
 from typing import List, Optional
 
-from sqlalchemy import ForeignKey, Integer, JSON, Numeric, String
+from sqlalchemy import ForeignKey, Integer, JSON, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -13,7 +12,7 @@ class Room(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(128), nullable=False)
     descriptions: Mapped[Optional[str]] = mapped_column(String(512))
-    price_per_day: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
+    price_per_day: Mapped[int] = mapped_column(Integer, nullable=False)
     services: Mapped[Optional[dict]] = mapped_column(JSON, default=dict)
     quality: Mapped[Optional[str]] = mapped_column(String(64))
     hotel_id: Mapped[int] = mapped_column(ForeignKey("hotels.id"), nullable=False)
